feat: optimize performance and support full-resolution image export
- Implement dual-path architecture: lightweight preview (/api/apply) with JPEG encoding vs full-resolution export (/api/export)
- Reduce backend memory usage: float64 → float32 pixel processing (2x bandwidth improvement)
- Add fast-path optimizations: early returns for neutral control values
- Implement sampled pixel analytics: 50k-60k point sampling for responsiveness on large images
- Add request cancellation with AbortController to prevent request backlog during slider drag
- Dynamic preview slider adjustment on image upload (automatically extends to uploaded image dimensions)
- Support 4K+ images with keep_resolution flag: download maintains original dimensions
- Frontend image dimension detection via readImageDimensions() for native width/height probing
- Normalized [0,1] pipeline: single entry/exit from normalized space reduces 255 conversions overhead

+++ src/la_core.py
@@ -9,7 +9,7 @@ from .filter_matrices import LUMINANCE_VECTOR
 
 def image_to_pixels(image: np.ndarray) -> np.ndarray:
     """Flatten HxWx3 image into Nx3 pixel matrix."""
-    return image.reshape(-1, 3).astype(np.float64)
+    return image.reshape(-1, 3).astype(np.float32, copy=False)
 
 
 def pixels_to_image(pixels: np.ndarray, shape: Tuple[int, int, int]) -> np.ndarray:
